requirements.txt
Flask 
selenium 
pandas
tabulate
python-telegram-bot==13.7
python-dotenv
requests
webdriver-manager
pyperclip
telethon
